% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistics.R
\name{AN}
\alias{AN}
\title{AN：宣告一個ANOVA物件}
\usage{
AN(list_of_vector, name)
}
\arguments{
\item{list_of_vector}{傳入一個含有多個向量的list。}

\item{name}{傳入一個字段。}
}
\value{
一個ANOVA物件。
}
\description{
快速，整潔的使用java中的Anova
}
\examples{
AN(list(c(1,2,3,4,5,6,7,8,9,10),c(1,2,3,4,5,6,7,8,9,10)),"test")
}
