% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistics.R
\name{DS}
\alias{DS}
\title{DS：宣告一個DescriptiveStatistics物件}
\usage{
DS(vector, name)
}
\arguments{
\item{vector}{傳入一個向量數列。}

\item{name}{傳入一個字段。}
}
\value{
一個DescriptiveStatistics物件。
}
\description{
快速，整潔的使用java中的DescriptiveStatistics
}
\examples{
DS(c(1,2,3,4,5,6,7,8,9,10),"test")
}
