% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistics.R
\name{HY}
\alias{HY}
\title{HY：宣告一個HypothesisTest物件}
\usage{
HY(DS)
}
\arguments{
\item{DS}{傳入一個DescriptiveStatistics物件。}
}
\value{
一個HypothesisTest物件。
}
\description{
以DescriptiveStatistics物件作為參數，快速，整潔的使用java中的HY
}
\examples{
HY(DS(c(1,2,3,4,5,6,7,8,9,10),"test"))
}
