% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistics.R
\name{LR}
\alias{LR}
\title{LR：宣告一個LinearRegression物件}
\usage{
LR(v1, v2, name)
}
\arguments{
\item{v1}{傳入一個向量數列作為因變量。}

\item{v2}{傳入一個向量數列作為自變量。}

\item{name}{傳入一個字段。}
}
\value{
一個LinearRegression物件。
}
\description{
快速，整潔的使用java中的LinearRegression
}
\examples{
LR(c(1,2,3,4,5,6,7,8,9,10),c(1,2,3,4,5,6,7,8,9,10),"test")
}
