% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistics.R
\name{LR_DS}
\alias{LR_DS}
\title{以DescriptiveStatistics物件作為參數，快速，整潔的使用java中的LinearRegression}
\usage{
LR_DS(DS, vector)
}
\arguments{
\item{DS}{傳入一個DescriptiveStatistics物件。}

\item{vector}{傳入一個向量數列作為自變量。}
}
\value{
一個LinearRegression物件。
}
\description{
以DescriptiveStatistics物件作為參數，快速，整潔的使用java中的LinearRegression
}
\examples{
LR_DS(DS(c(1,2,3,4,5,6,7,8,9,10),"test"),c(1,2,3,4,5,6,7,8,9,10))
}
