% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DescriptiveStatistics.R
\name{LR_predict}
\alias{LR_predict}
\title{LR_predict：使用LinearRegression物件進行預測}
\usage{
LR_predict(lr, x)
}
\arguments{
\item{lr}{傳入一個LinearRegression物件。}

\item{x}{傳入一個數字預測。}
}
\value{
一個預測值。
}
\description{
使用LinearRegression中的method predict進行預測
}
\examples{
LR_predict(LR(c(1,2,3,4,5,6,7,8,9,10),c(1,2,3,4,5,6,7,8,9,10),"test"),1)
}
